Insert_Buffer_Properties.bsh is UTF-8 - bug #3311234
Macros can't be assumed to be all UTF-8.
They have always been read using platform's default encoding
so it's not changed to not break user's macros.

The only core macro with non ascii is Insert_Buffer_Properties.bsh
I add a buffer-local-property encoding=UTF-8 to remove a warning
on windows, where it's not UTF-8 by default, but it works only
for people having the autodetector active.

git-svn-id: svn+ssh://svn.code.sf.net/p/jedit/svn/jEdit/trunk@22807 6b1eeb88-9816-0410-afa2-b43733a0f04e